# docs/swagger.yaml
basePath: /api
definitions:
...
  apierror.ApiError:
    properties:
      code:
        type: integer
      error: {}
    type: object
...
  models.CreateAttemptRequest:
    properties:
      assignment_id:
        type: integer
      code:
        type: string
      lang:
        type: string
      task_id:
        type: integer
    type: object
...
  models.GetUserTaskAttemptsResponse:
    properties:
      assignment_id:
        type: integer
      code:
        type: string
      id:
        type: integer
      lang:
        type: string
      memory_usage:
        type: integer
      running_time:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:4000
info:
  contact: {}
  title: Rankode
  version: "1.0"
paths:
...
  /attempts:
    get:
      consumes:
      - application/json
      description: Retrieves all attempts submitted by the authenticated user for
        a given task.
      parameters:
      - description: Task ID
        in: query
        name: taskId
        required: true
        type: integer
      - description: Assignment ID
        in: query
        name: assignmentId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of attempts
          schema:
            items:
              $ref: '#/definitions/models.GetUserTaskAttemptsResponse'
            type: array
        "400":
          description: Bad request (e.g., missing or invalid taskId)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuoth: []
      summary: List user’s attempts for a specific task
      tags:
      - Attempts
    post:
      consumes:
      - application/json
      description: Submits a new code attempt for execution against a task’s test
        cases.
      parameters:
      - description: Attempt creation payload
        in: body
        name: attempt
        required: true
        schema:
          $ref: '#/definitions/models.CreateAttemptRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Attempt queued successfully
        "400":
          description: Bad request (e.g., still running or invalid input)
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/apierror.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apierror.ApiError'
      security:
      - ApiKeyAuth: []
      summary: Create a new attempt
      tags:
      - Attempts
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"